refactor: create pages and components
